requirements.txt
Flask
google-generativeai>=0.7.0
gunicorn
requests